Clean up using directives and minor code hygiene fixes
Removed unused or redundant using directives across the codebase, especially in test and implementation files. Added missing usings where required for compilation or clarity. Reordered and grouped usings for consistency. Made minor code corrections, such as removing unnecessary type casts and fixing parameter types. Updated dictionary initializations in test data for consistent formatting. These changes improve code hygiene, reduce clutter, and ensure only necessary dependencies are included.

+++ src/TotallyHotArcRouter.Gui/Components/ProvidersAdmin.razor.cs
@@ -430,7 +430,7 @@ public partial class ProvidersAdmin
                 var row = usage.Rows.FirstOrDefault(r => r.UsageDay == day && r.Model == model);
                 return row is null
                     ? 0m
-                    : (decimal)(row.InputTokens + row.OutputTokens + row.CacheCreationTokens + row.CacheReadTokens);
+                    : row.InputTokens + row.OutputTokens + row.CacheCreationTokens + row.CacheReadTokens;
             }).ToList();
             return new DistributionSeries(model, ReportedUsageSeriesColors[index % ReportedUsageSeriesColors.Length], data);
         }).ToList();
